flask/requirements.txt
Flask
torch
torchvision
albumentations==0.4.6
pillow
